H2 database and real implementation of Card stuff

+++ src/main/java/microservice/services/CardService.java
@@ -16,7 +16,7 @@ public class CardService {
         this.cardRepository = cardRepository;
     }
 
-    public List<Card> getCards(List<Integer> cardIdList) {
+    public List<Card> getCards(List<Long> cardIdList) {
         return cardRepository.getCards(cardIdList);
     }
 }
